feat: update assembly to allow for non-simd execution
